Fix number of temp sensors to 7; Change datatype to int.
They should be interpreted as temp/100.

+++ src/include/protocol_sensors.h
@@ -60,7 +60,7 @@ struct SensorIMU {
 } __attribute__((packed));
 
 struct SensorTemp {
-    float temp;
+    int temp;
 }__attribute__((packed));
 
 struct SensorADCS {
